vmm: igvm: populate supported extended state CPUID subleaves
The CPUID page in the IGVM file will no longer have hardcoded entries
for processor extended state CPUID subleaves. Instead, it will have
entries only for the 0x0 and 0x1 subleaves and the VMM should query and
import the supported subleaves.

Hardcoding some unsupported leaves with zeroes can sometimes lead to the
PSP rejecting the CPUID page. One instance we've observed of this is
when AVX512 is supported by the hardware the PSP doesn't allow setting
zeroes for the corresponding subleaves even if AVX512 is not exposed to
the VM. However, the PSP is okay if we just omit those subleaves in our
CPUID page. This behavior is seen in the newer AMD Genoa hardware where
AVX512 is supported.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>